+++ CHANGELOG.md
@@ -18,6 +18,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `CharSequence`
     - `hasLengthEqualTo` / `hasLengthNotEqualTo`  
     - `isBlank` / `isNotBlank`  
+    - `isMatching` / `isNotMatching`
 
 ### Changed